30.11.21

+++ src/main/java/com/kon/EShop/util/FileManager.java
@@ -1,11 +1,17 @@
 package com.kon.EShop.util;
 
+import com.kon.EShop.DelLabel;
 import com.kon.EShop.model.MyFile;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.util.ResourceUtils;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.imageio.ImageIO;
+import javax.swing.*;
+import java.awt.*;
+import java.awt.geom.Rectangle2D;
+import java.awt.image.BufferedImage;
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
@@ -30,19 +36,22 @@ public class FileManager {
 
     public String upload(MultipartFile resource, String keyName, String where) throws IOException {
         Files.createDirectories(Paths.get(where));
+
         Path file = Files.createFile(Paths.get(where, keyName));
         resource.transferTo(file);
         return file.toString();
     }
 
     public void delete(String key, String where) throws IOException {
-        Path path = Paths.get(this.path + (where.equals("label") ?
-            label :
-            where.equals("csv") ?
-                getAbsolutePath() + "static/csv/" :
-                photo) + key);
+        if (!key.equals("temp.jpg") && !key.equals("favicon.ico")) {
+            Path path = Paths.get(this.path + (where.equals("label") ?
+                    label :
+                    where.equals("csv") ?
+                            getAbsolutePath() + "static/csv/" :
+                            photo) + key);
 
-        Files.delete(path);
+            Files.delete(path);
+        }
     }
 
     public List<String> doUpload(MultipartFile[] file, String dir) throws IOException {
@@ -50,7 +59,10 @@ public class FileManager {
         for (MultipartFile fileData : file) {
             String name = fileData.getOriginalFilename();
             if (name != null && name.length() > 0) {
-                upload(fileData, name, path + ((dir.equals("label")) ? label : photo));
+                if (dir.equals("label"))
+                    upload(fileData, name, path + label);
+                if (dir.equals("photos"))
+                    waterMark("AutoMir", path + photo, fileData);
                 uploaded.add(name);
             }
         }
@@ -69,7 +81,7 @@ public class FileManager {
     }
 
     public static String getPath(String where) {
-        String csv = where.equals("csv") ? getAbsolutePath() + "static/csv/" : "static/img/" + where + "/";
+        String csv = where.equals("csv") ? getAbsolutePath() + "static/csv/" : "static/image/" + where + "/";
         return csv.substring(6);
     }
 
@@ -98,6 +110,37 @@ public class FileManager {
             e.printStackTrace();
         }
         return myFile;
+    }
+
+    public void delLabel(DelLabel entity) throws IOException {
+        if (entity != null && entity.getLabel() != null) {
+            if (!entity.getLabel().equals("temp.jpg"))
+                this.delete(entity.getLabel(), "label");
+        }
+    }
+
+    public void waterMark(String mark, String path, MultipartFile file) throws IOException {
+        ImageIcon image = new ImageIcon(file.getBytes());
+        BufferedImage bufferedImage = new BufferedImage(image.getIconWidth(), image.getIconHeight(), BufferedImage.TYPE_INT_BGR);
+        Graphics2D g2d = (Graphics2D) bufferedImage.getGraphics();
+        g2d.drawImage(image.getImage(), 0, 0, null);
+        AlphaComposite alpha = AlphaComposite.getInstance(AlphaComposite.SRC_OVER, 0.5f);
+        g2d.setComposite(alpha);
+        g2d.setColor(Color.WHITE);
+        g2d.setRenderingHint(RenderingHints.KEY_TEXT_ANTIALIASING, RenderingHints.VALUE_TEXT_ANTIALIAS_ON);
+        g2d.setFont(new Font("Arial", Font.BOLD, 60));
+
+        FontMetrics fontMetrics = g2d.getFontMetrics();
+        Rectangle2D rect = fontMetrics.getStringBounds(mark, g2d);
+        g2d.drawString(mark, (image.getIconWidth() - (int) rect.getWidth()) / 2, (image.getIconHeight() - (int) rect.getHeight()) / 2);
+        g2d.drawString(mark, (image.getIconWidth() - (int) rect.getWidth()) / 3, (image.getIconHeight() - (int) rect.getHeight()) / 3);
+
+        g2d.dispose();
+
+        String name = file.getOriginalFilename();
+        OutputStream out = new PrintStream(path + name);
+        ImageIO.write(bufferedImage, "jpg", out);
+        out.close();
     }
 
 }
